Working MG Katz with nstart plus other edits
