+++ python/cugraph/cugraph/tests/mg/test_mg_katz_centrality.py
@@ -31,8 +31,7 @@ def setup_function():
     gc.collect()
 
 
-IS_DIRECTED = [True]
-#IS_DIRECTED = [True, False]
+IS_DIRECTED = [True, False]
 
 
 @pytest.mark.skipif(
@@ -101,7 +100,7 @@ def test_dask_katz_centrality(dask_client, directed):
     is_single_gpu(), reason="skipping MG testing on Single GPU system"
 )
 @pytest.mark.parametrize("directed", IS_DIRECTED)
-def test_dask_katz_centrality_multi_column(dask_client, directed):
+def test_dask_katz_centrality_nstart(dask_client, directed):
     input_data_path = (RAPIDS_DATASET_ROOT_DIR_PATH /
                        "karate.csv").as_posix()
     print(f"dataset={input_data_path}")
@@ -123,17 +122,30 @@ def test_dask_katz_centrality_multi_column(dask_client, directed):
     largest_out_degree = largest_out_degree["degree"].iloc[0]
     katz_alpha = 1 / (largest_out_degree + 1)
 
-    #dg.view_edge_list.compute()
-
     mg_res = dcg.katz_centrality(dg, alpha=katz_alpha, max_iter=50, tol=1e-6)
-    #breakpoint()
     mg_res = mg_res.compute()
 
-    #estimate = dask_cudf.from_cudf(cudf.DataFrame(), npartitions=2)
     estimate = mg_res.copy()
-    estimate = estimate.rename(columns={"vertex": "vertex", "katz_centrality": "values"})
+    estimate = estimate.rename(columns={"vertex": "vertex",
+                                        "katz_centrality": "values"})
     estimate["values"] = 0.5
-    mg_final_res = dcg.katz_centrality2(dg, alpha=katz_alpha, nstart=estimate,
-                                       max_iter=50, tol=1e-6)
-    mg_final_res = mg_final_res.compute()
-    print(mg_final_res)
+
+    mg_estimate_res = dcg.katz_centrality(dg, alpha=katz_alpha,
+                                           nstart=estimate,
+                                           max_iter=50, tol=1e-6)
+    mg_estimate_res = mg_estimate_res.compute()
+
+    err = 0
+    tol = 1.0e-05
+    compare_res = mg_res.merge(
+        mg_estimate_res, on="vertex", suffixes=["_dask", "_nstart"]
+    )
+
+    for i in range(len(compare_res)):
+        diff = abs(
+            compare_res["katz_centrality_dask"].iloc[i]
+            - compare_res["katz_centrality_nstart"].iloc[i]
+        )
+        if diff > tol * 1.1:
+            err = err + 1
+    assert err == 0
