+++ python/cugraph/cugraph/tests/test_hits.py
@@ -107,7 +107,7 @@ def test_hits(benchmark, input_expected_output):
     cugraph_hits["nx_hubs"] = cudf.Series.from_pandas(pdf[0])
     pdf = pd.DataFrame.from_dict(nx_authorities, orient="index").sort_index()
     cugraph_hits["nx_authorities"] = cudf.Series.from_pandas(pdf[0])
-    #breakpoint()
+
     hubs_diffs1 = cugraph_hits.query('hubs - nx_hubs > 0.00001')
     hubs_diffs2 = cugraph_hits.query('hubs - nx_hubs < -0.00001')
     authorities_diffs1 = cugraph_hits.query(
